Move files related to quest tab to a separate folder
